fixed getting properties and reorganized code #4
